pyproject.toml
[tool.poetry]
name = "th-tasks"
version = "0.1.0"
description = ""
authors = ["DimitryRukhadze <lokiinflames@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0"
django-ninja = "^1.1.0"
django-extensions = "^3.2.3"
pytest = "^7.4.4"
pytest-django = "^4.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
